+++ R/format_sumstats_for_magma.r
@@ -21,6 +21,7 @@ format_sumstats_for_magma <- function(path) {
     mungesumstats_deprecation_msg()
 }
 
+
 format.sumstats.for.magma <- function(path){
     format_sumstats_for_magma(path)
 }