+++ R/list_remote_files.R
@@ -13,7 +13,9 @@
 #' @importFrom XML htmlParse xpathSApply free
 list_remote_files <- function(URL,
                               pattern = "*.zip") {
-    # URL <- "https://ctg.cncr.nl/software/MAGMA/prog/"
+ 
+    # doc <- rvest::read_html("https://vu.data.surfsara.nl/index.php/s/8qDPUbOTrZ9lW2b/")
+    # rvest::html_elements(doc,"a")
     doc <- XML::htmlParse(readLines(URL), asText = TRUE)
     links <- XML::xpathSApply(doc, "//a/@href")
     XML::free(doc)