+++ R/magma_download_binary.R
@@ -9,10 +9,10 @@
 #' @keywords internal 
 magma_download_binary <- function(magma_url,
                                   dest_dir = find_install_dir(),
+                                  destfile = basename(magma_url),
                                   verbose = TRUE) {
-    #### Download the appropriate executable ####
-    destfile <- file.path(dest_dir, basename(magma_url))
-    destfile <- fix_path(destfile) 
+    #### Download the appropriate executable #### 
+    destfile <- fix_path(file.path(dest_dir,destfile)) 
     messager("Downloading MAGMA executable ==>",dest_dir, v = verbose)
     # Create the decompressed file name in advance
     destpath <- gsub("\\.zip$", "", destfile)