+++ tests/testthat/test-magma_links.R
@@ -1,30 +1,27 @@
 test_that("magma_links works", {
     
     if(!is_32bit()){
-        to_numeric <- function(x){
-            as.numeric(
-                gsub(
-                    paste( c(letters,".zip"), collapse = "|" ),"",
-                    stringr::str_split(basename(x),"_")[[1]][2]
-                    )
-                )
+        to_version <- function(x){
+          package_version(
+            gsub("v","",MAGMA.Celltyping:::magma_links_versions(x))
+          )
         }
         
         #### Mac ####
-        magma_mac <- MAGMA.Celltyping:::magma_links(latest_only = TRUE,
+        magma_mac <-MAGMA.Celltyping:::magma_links(latest_only = TRUE,
                                os = "Mac",
                                verbose = FALSE)
-        testthat::expect_gte(to_numeric(magma_mac), 1.1) 
+        testthat::expect_true(to_version(magma_mac)>=1.1) 
         
         #### Windows ####
         magma_win <- MAGMA.Celltyping:::magma_links(latest_only = TRUE,
                                                     os = "Windows")
-        testthat::expect_gte(to_numeric(magma_win), 1.1) 
+        testthat::expect_true(to_version(magma_win)>=1.1) 
         
         #### Linux ####
         magma_linux <- MAGMA.Celltyping:::magma_links(latest_only = TRUE,
                                                       os = "Linux")
-        testthat::expect_gte(to_numeric(magma_linux), 1.1) 
+        testthat::expect_true(to_version(magma_linux)>=1.1) 
           
         #### All ####
         meta <- MAGMA.Celltyping:::magma_links(latest_only = FALSE,